Stability: Split Envoy Service into Reusable Parts (#8365)
* split envoy.go in sub-files

* updat references to envoyFactory

* bad find/replace

* simplify interface

* remove import cycle

* export dockeroptions, inclue Run() method to support ee implementations

* include AccessLogs to mirror enterprise

* expose StructureConfigDump method on EnvoyInstance

* add changelog

* fix find/replace, update env comment, remove uncessary file

* don't export bootstrap builder

* make -B update-licenses

* push updates per PR review, collapse instance_manager into factory

* don't export factoryimpl

---------
