+++ docs/content/static/content/osa_provided.md
@@ -36,7 +36,6 @@ Name|Version|License
 [rotisserie/eris](https://github.com/rotisserie/eris)|v0.4.0|MIT License
 [saiskee/gettercheck](https://github.com/saiskee/gettercheck)|v0.0.0-20210820204958-38443d06ebe0|MIT License
 [sergi/go-diff](https://github.com/sergi/go-diff)|v1.1.0|MIT License
-[gopsutil/v3](https://github.com/shirou/gopsutil)|v3.23.4|BSD 3-clause "New" or "Revised" License
 [spf13/afero](https://github.com/spf13/afero)|v1.9.2|Apache License 2.0
 [spf13/cobra](https://github.com/spf13/cobra)|v1.6.1|Apache License 2.0
 [spf13/pflag](https://github.com/spf13/pflag)|v1.0.5|BSD 3-clause "New" or "Revised" License
