+++ test/e2e/e2e_suite_test.go
@@ -29,7 +29,7 @@ func TestE2e(t *testing.T) {
 }
 
 var (
-	envoyFactory  *services.EnvoyFactory
+	envoyFactory  services.EnvoyFactory
 	consulFactory *services.ConsulFactory
 	vaultFactory  *services.VaultFactory
 
@@ -40,8 +40,8 @@ var (
 
 var _ = BeforeSuite(func() {
 	var err error
-	envoyFactory, err = services.NewEnvoyFactory()
-	Expect(err).NotTo(HaveOccurred())
+	envoyFactory = services.MustEnvoyFactory()
+
 	consulFactory, err = services.NewConsulFactory()
 	Expect(err).NotTo(HaveOccurred())
 	vaultFactory, err = services.NewVaultFactory()
@@ -55,7 +55,7 @@ var _ = BeforeSuite(func() {
 })
 
 var _ = AfterSuite(func() {
-	_ = envoyFactory.Clean()
+	envoyFactory.MustClean()
 	_ = consulFactory.Clean()
 	_ = vaultFactory.Clean()
 })
